+++ src/test/java/org/folio/entitlement/integration/kafka/CapabilitiesModuleEventPublisherTest.java
@@ -12,6 +12,8 @@ import static org.folio.entitlement.domain.model.CommonStageContext.PARAM_TENANT
 import static org.folio.entitlement.domain.model.ModuleStageContext.PARAM_INSTALLED_MODULE_DESCRIPTOR;
 import static org.folio.entitlement.domain.model.ModuleStageContext.PARAM_MODULE_DESCRIPTOR;
 import static org.folio.entitlement.domain.model.ModuleStageContext.PARAM_MODULE_TYPE;
+import static org.folio.entitlement.integration.kafka.KafkaEventUtils.CAPABILITIES_TOPIC;
+import static org.folio.entitlement.integration.kafka.KafkaEventUtils.TOPIC_TENANT_COLLECTION_KEY;
 import static org.folio.entitlement.integration.kafka.model.ModuleType.MODULE;
 import static org.folio.entitlement.integration.kafka.model.ModuleType.UI_MODULE;
 import static org.folio.entitlement.support.TestConstants.APPLICATION_FLOW_ID;
@@ -27,9 +29,12 @@ import static org.folio.entitlement.support.TestUtils.readCapabilityEvent;
 import static org.folio.entitlement.support.TestUtils.readModuleDescriptor;
 import static org.folio.entitlement.support.TestValues.moduleFlowParameters;
 import static org.folio.entitlement.support.TestValues.moduleStageContext;
+import static org.folio.integration.kafka.producer.KafkaUtils.getTenantTopicName;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 
@@ -117,6 +122,7 @@ class CapabilitiesModuleEventPublisherTest {
     var topicName = capabilitiesTenantCollectionTopic();
     doNothing().when(kafkaEventPublisher).send(eq(topicName), messageKeyCaptor.capture(), eventCaptor.capture());
     when(tenantEntitlementKafkaProperties.isProducerTenantCollection()).thenReturn(true);
+    when(tenantEntitlementKafkaProperties.getTenantCollectionQualifier()).thenReturn(TOPIC_TENANT_COLLECTION_KEY);
 
     moduleEventPublisher.execute(stageContext);
 
@@ -231,10 +237,21 @@ class CapabilitiesModuleEventPublisherTest {
   }
 
   @Test
-  void getTopicNameByTenantCollection_positive() {
-    var actual = moduleEventPublisher.getTopicNameByTenantCollection();
+  void execute_positive_customTenantCollectionTopic() {
+    var request = entitlementRequest();
+    var descriptor = readModuleDescriptor("json/events/capabilities/be-module-desc.json");
+    var contextData = Map.of(PARAM_TENANT_NAME, TENANT_NAME);
+    var flowParameters = moduleFlowParameters(request, descriptor);
+    var stageContext = moduleStageContext(FLOW_STAGE_ID, flowParameters, contextData);
+    stageContext.withStageId(UUID.randomUUID());
 
-    assertThat(actual).isEqualTo("folio.ALL.mgr-tenant-entitlements.capability");
+    when(tenantEntitlementKafkaProperties.isProducerTenantCollection()).thenReturn(true);
+    when(tenantEntitlementKafkaProperties.getTenantCollectionQualifier()).thenReturn("COLLECTIONA");
+
+    moduleEventPublisher.execute(stageContext);
+
+    var topicName = getTenantTopicName(CAPABILITIES_TOPIC, "COLLECTIONA");
+    verify(kafkaEventPublisher).send(eq(topicName), eq(TENANT_NAME), any());
   }
 
   private static EntitlementRequest entitlementRequest() {